// apps/web/src/lib/api/generated/models/appraisalAnswerSchemaDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */
import type { AppraisalAnswerOptionDto } from './appraisalAnswerOptionDto.ts';

export type AppraisalAnswerSchemaDto =
	| {
			kind: 'enum';
			options: AppraisalAnswerOptionDto[];
	  }
	| {
			kind: 'boolean';
	  }
	| {
			kind: 'scale';
			labels: { [key: string]: string };
			max: number;
			min: number;
	  }
	| {
			kind: 'text';
			/** @minimum 0 */
			max_length: number;
	  };